Remove the concept of `dry_run`
The flag was false by default, adds complexity to the code and can be
confused with the other disabled states `use_global_uid`, the
global_uid_options[:disabled] and `global_uid_disabled` on the model.

Use `global_uid_options[:disabled]` if you want to disable the
functionality. There's a performance spec for benchmarking.