Remove ServerVariables module
This module adds confusion to the implementation, one could mistakenly
think it's how to configure a production database, instead,
remove it and rely on the database.yml to have the variables set on
in the dev and test environment for this gem.

For production, clients should be setting the values directly in their
MySQL configuration.
https://dev.mysql.com/doc/refman/5.6/en/replication-options-master.html#sysvar_auto_increment_increment

The `offset` variables were also unused and have been removed