Add performance spec
This spec highlights the overhead created by using GlobalUID. The
results will vary per run but it will give you an idea and highlight
performance issues if they're introduced.

Output from my local machine:

```
Warming up --------------------------------------
       WithGlobalUID    77.000  i/100ms
    WithoutGlobalUID   127.000  i/100ms
Calculating -------------------------------------
       WithGlobalUID    923.498  (± 8.3%) i/s -      4.620k in   5.041687s
    WithoutGlobalUID      1.330k (± 6.5%) i/s -      6.731k in   5.081601s

Comparison:
    WithoutGlobalUID:     1330.4 i/s
       WithGlobalUID:      923.5 i/s - 1.44x  slower
```

As you can see, the performance overhead is noticable but minimal. With
GlobalUID we're able to generate around 923 records per second. This is
not indicative of the performance you'll get in production, where the
databases are more powerful and there's network latency.

The assertion has been commented out due to a lot of variance between
runs, things I've tried:
 * Increase the warm-up (up to 5 sec) and time intervals (up to 20 sec)
 * Use `kalibera` with `benchmark-ips` and play with confidence levels
 * Increase the threshold to 2 seconds, it passes but it's not really
 valuable
 * Run many times and check an average or check that one has the result
 we're after, caters for outliers but also not valuable. Will miss
 upward trends

We could look at commiting the results and keeping track in source
control, that would give us visibility when/if performance issues are
introduced but I'm parking that decision for another day.