Remove unreachable `NoServersAvailableException` references
Call the notifier when there's a `NoServersAvailableException` exception

----

Removal of:
`raise NoServersAvailableException if servers.empty?`

If the servers are empty, the connections will also be empty, resulting
in an exception further down, e.g. `if get_connections.empty?`

----

Removal of:
`raise NoServersAvailableException, "All global UID servers are gone!"`

Unreachable, if `with_servers` is unable to perform the operation on any
of the `id_servers` an exception is thrown, preventing us ever getting
to this line. Remove it so it doesn't mislead others.

----

Removal of:
`raise "You haven't configured any id servers" if id_servers.nil? or id_servers.empty?`

A `NoServersAvailableException` will be raised instead

----

Removal of:
`servers.each { |server| server.update_retry_at(0) }`

We can remove this line, it was added in the initial commit and was only
useful when the connection was shared between processes. It was a no-op
once the gem was made thread safe in #21

The spec passed due to the explicit resque