Fix validation error when working with multiple models.
Each model using global_uids has an ID table created for it on the
allocation server. If we're keeping track of recently allocated IDs, we
need to scope it by table name as the sequence between tables will be
different.

To resolve this, the Server now contains a hash of Allocators, keyed by
table name. The Server will validate the `increment_by` value on the DB
during connection initialization, and whenever an allocation is made
we'll create or retrieve an allocator for that table, using validated
connection reference.

The connection validation was moved to reduce overhead, performing
it per table, per process would impact performance and as a result of
this, some of the specs have been moved around.